Extract shared query utils into dedicated module
