+++ agents/security_advisories/lambda/dsl_query_builder.py
@@ -9,91 +9,24 @@ for vulnerability queries. It replaces the previous agentic search flow that
 relied on an ML-powered pipeline for NL→DSL translation.
 
 Functions:
-    resolve_version_tag: Map user-provided version to canonical tag format
+    resolve_version_tag: Map user-provided version to canonical tag format (re-exported from query_utils)
     query_vulnerabilities: Construct and execute a DSL query for vulnerability scans
     query_advisories: Query advisories index to filter CVEs by age and/or severity
 """
 
 import json
 import logging
-import re
 from datetime import datetime, timedelta, timezone
 from typing import Any, Dict, List, Optional, Set, Tuple
 
-import semver
 from aws_utils import get_latest_scans_index, opensearch_request
+from query_utils import connection_error, error_response, resolve_version_tag
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 # Default query size — matches the previous agentic search configuration
 _DEFAULT_QUERY_SIZE = 1000
-
-# Strict two-part numeric pattern (e.g., "3.7", "2.19") to avoid
-# misclassifying pre-release/build metadata strings like "3.7-rc".
-_TWO_PART_RE = re.compile(r'^\d+\.\d+$')
-
-
-def _classify_version(version: str) -> str:
-    """Classify a version string into a known category for dispatch.
-
-    Returns one of: 'origin_prefixed', 'main_alias', 'three_part', 'two_part', 'unknown'.
-    """
-    if version.startswith('origin/'):
-        return 'origin_prefixed'
-    if version.lower() in ('main', 'latest'):
-        return 'main_alias'
-    try:
-        semver.Version.parse(version)
-        return 'three_part'
-    except (ValueError, TypeError):
-        pass
-    if _TWO_PART_RE.match(version):
-        try:
-            semver.Version.parse(f'{version}.0')
-            return 'two_part'
-        except (ValueError, TypeError):
-            pass
-    return 'unknown'
-
-
-def resolve_version_tag(version: str) -> str:
-    """Map a user-provided version string to the canonical project.tag format.
-
-    The scans index stores release branch tags as ``origin/{major}.{minor}``
-    and specific release version tags as three-part semver (e.g., ``2.19.6``).
-
-    Mapping rules:
-      - Already prefixed with ``"origin/"`` → returned as-is
-      - ``"main"`` or ``"latest"`` → ``"origin/main"``
-      - Two-part version (e.g., ``"3.7"``) → ``"origin/3.7"`` (branch tag)
-      - Three-part version (e.g., ``"3.7.0"``, ``"2.19.6"``) → ``"origin/3.7"``, ``"origin/2.19"`` (branch tag)
-      - Non-parseable input → returned as-is (for exact tag lookups)
-
-    Args:
-        version: User-provided version or tag string.
-
-    Returns:
-        The canonical tag string to use in queries.
-    """
-    if not version:
-        return version
-
-    match _classify_version(version):
-        case 'origin_prefixed':
-            resolved = version
-        case 'main_alias':
-            resolved = 'origin/main'
-        case 'three_part':
-            parsed = semver.Version.parse(version)
-            resolved = f'origin/{parsed.major}.{parsed.minor}'
-        case 'two_part':
-            resolved = f'origin/{version}'
-        case _:
-            resolved = version
-
-    logger.info(f"RESOLVE_TAG: '{version}' -> '{resolved}'")
-    return resolved
 
 
 def _build_dsl_query(
@@ -189,54 +122,6 @@ def _execute_query(index: str, query_body: str) -> Dict[str, Any]:
     return result
 
 
-def _error_response(
-    error_type: str,
-    message: str,
-    status_code: Optional[int] = None,
-) -> Dict[str, Any]:
-    """Return a consistent error response dict.
-
-    Args:
-        error_type: Category of the error (e.g., opensearch_error, connection_error).
-        message: Human-readable error description.
-        status_code: Optional HTTP status code from OpenSearch.
-
-    Returns:
-        Error dict with status, type, retryable, message, and optional status_code.
-    """
-    error = {
-        'status': 'error',
-        'type': error_type,
-        'retryable': False,
-        'message': message,
-    }
-
-    if status_code is not None:
-        error['status_code'] = status_code
-
-    return error
-
-
-def _connection_error(exception: Exception) -> Dict[str, Any]:
-    """Return a connection error without leaking internal details.
-
-    Logs the original exception internally for diagnostics, then returns
-    a sanitized error dict via _error_response for consistent structure.
-
-    Args:
-        exception: The caught exception from the connection failure.
-
-    Returns:
-        Sanitized error dict with consistent structure.
-    """
-    logger.error(f'CONNECTION_ERROR: {type(exception).__name__}: {exception}')
-    return _error_response(
-        'connection_error',
-        'Failed to connect to the OpenSearch cluster. '
-        'The service may be temporarily unavailable.',
-    )
-
-
 def query_vulnerabilities(
     version: Optional[str] = None,
     project_name: Optional[str] = None,
@@ -262,7 +147,7 @@ def query_vulnerabilities(
             f'SECURITY_ADVISORIES_DSL_QUERY_FAILED: '
             f'Could not resolve scans index: {e}',
         )
-        return _error_response('index_resolution_error', str(e))
+        return error_response('index_resolution_error', str(e))
 
     # Resolve version tag:
     # - version provided → resolve to canonical tag format
@@ -303,7 +188,7 @@ def query_vulnerabilities(
             logger.error(
                 f'SECURITY_ADVISORIES_DSL_QUERY_FAILED: {error_msg}',
             )
-            return _error_response(
+            return error_response(
                 'opensearch_error',
                 f'OpenSearch query failed: {error_msg}',
                 status_code=status_code,
@@ -313,7 +198,7 @@ def query_vulnerabilities(
         logger.error(
             f'SECURITY_ADVISORIES_DSL_QUERY_FAILED: {error_msg}',
         )
-        return _connection_error(e)
+        return connection_error(e)
 
     return result
